+++ pyopenvidu/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Marcell Pünkösd"""
 __email__ = 'punkosdmarcell@rocketmail.com'
-__version__ = '0.2.0'
+__version__ = '0.2.1'
 
 from .openvidu import OpenVidu